adding backend

+++ dating_site/dating_site/admin.aspx.cs
@@ -40,63 +40,66 @@ namespace dating_site
 
         protected void Button1_Click(object sender, EventArgs e)
         {
-
-            string filtered_email = erase_email.Text.ToString();
-            int index = filtered_email.IndexOf('@');
-            filtered_email = filtered_email.Substring(0, index);
-
-            //queries
-            string checkemail = "select email from usertable WHERE email LIKE '" + filtered_email + "%'";
-            string checkpassword = "select password from usertable WHERE email LIKE '" + filtered_email + "%'";
-
-
-
-            SqlConnection connection = new SqlConnection("Data Source = uvuserdata.mssql.somee.com; Initial Catalog = uvuserdata; Persist Security Info = True; User ID = yuvrajvirk55_SQLLogin_1; Password = nm6ecevlt8");
-
-            SqlCommand command_email = new SqlCommand(checkemail, connection);
-            SqlCommand command_password = new SqlCommand(checkpassword, connection);
-
-            connection.Open();
-
-            //database values
-            string db_email = (string)command_email.ExecuteScalar();
-            string db_password = (string)command_password.ExecuteScalar();
-
-            db_password = Encrypt.DecryptString(db_password, erase_password.Text);
-
-
-            if ((!string.IsNullOrEmpty(db_email) & !string.IsNullOrEmpty(db_password)) && (db_email.Equals(erase_email.Text.ToString()) & db_password.Equals(erase_password.Text.ToString())))
+            if (erase_email.Text != null && erase_password.Text != null)
             {
 
-                Page.ClientScript.RegisterClientScriptBlock(typeof(Page), "Alert", "alert('uv !')", true);
+                string filtered_email = erase_email.Text.ToString();
+                int index = filtered_email.IndexOf('@');
+                filtered_email = filtered_email.Substring(0, index);
 
+                //queries
+                string checkemail = "select email from usertable WHERE email LIKE '" + filtered_email + "%'";
+                string checkpassword = "select password from usertable WHERE email LIKE '" + filtered_email + "%'";
+
+
+
+                SqlConnection connection = new SqlConnection("Data Source = uvuserdata.mssql.somee.com; Initial Catalog = uvuserdata; Persist Security Info = True; User ID = yuvrajvirk55_SQLLogin_1; Password = nm6ecevlt8");
+
+                SqlCommand command_email = new SqlCommand(checkemail, connection);
+                SqlCommand command_password = new SqlCommand(checkpassword, connection);
+
+                connection.Open();
+
+                //database values
+                string db_email = (string)command_email.ExecuteScalar();
+                string db_password = (string)command_password.ExecuteScalar();
+
+                db_password = Encrypt.DecryptString(db_password, erase_password.Text);
+
+
+                if ((!string.IsNullOrEmpty(db_email) & !string.IsNullOrEmpty(db_password)) && (db_email.Equals(erase_email.Text.ToString()) & db_password.Equals(erase_password.Text.ToString())))
+                {
+
+                    Page.ClientScript.RegisterClientScriptBlock(typeof(Page), "Alert", "alert('uv !')", true);
+
+                }
+
+                else
+                    Page.ClientScript.RegisterClientScriptBlock(typeof(Page), "Alert", "alert('Incorrect credentials !')", true);
+
+                /*
+                    System.Data.SqlClient.SqlConnection sqlConnection1 = new System.Data.SqlClient.SqlConnection("Data Source = uvuserdata.mssql.somee.com; Initial Catalog = uvuserdata; Persist Security Info = True; User ID = yuvrajvirk55_SQLLogin_1; Password = nm6ecevlt8");
+                    System.Data.SqlClient.SqlCommand cmd = new System.Data.SqlClient.SqlCommand();
+
+                    cmd.CommandType = System.Data.CommandType.Text;
+                    cmd.CommandText = "delete from usertable";
+                    cmd.Connection = sqlConnection1;
+
+
+                    sqlConnection1.Open();
+                    cmd.ExecuteNonQuery();
+
+                    cmd.CommandType = System.Data.CommandType.Text;
+                    cmd.CommandText = "delete from userinterest";
+                    cmd.Connection = sqlConnection1;
+
+                    cmd.ExecuteNonQuery();
+                    sqlConnection1.Close();*/
             }
-
             else
-                Page.ClientScript.RegisterClientScriptBlock(typeof(Page), "Alert", "alert('Incorrect credentials !')", true);
+                Page.ClientScript.RegisterClientScriptBlock(typeof(Page), "Alert", "alert('Enter credientials !')", true);
 
-            /*
-                System.Data.SqlClient.SqlConnection sqlConnection1 = new System.Data.SqlClient.SqlConnection("Data Source = uvuserdata.mssql.somee.com; Initial Catalog = uvuserdata; Persist Security Info = True; User ID = yuvrajvirk55_SQLLogin_1; Password = nm6ecevlt8");
-                System.Data.SqlClient.SqlCommand cmd = new System.Data.SqlClient.SqlCommand();
-
-                cmd.CommandType = System.Data.CommandType.Text;
-                cmd.CommandText = "delete from usertable";
-                cmd.Connection = sqlConnection1;
-
-
-                sqlConnection1.Open();
-                cmd.ExecuteNonQuery();
-
-                cmd.CommandType = System.Data.CommandType.Text;
-                cmd.CommandText = "delete from userinterest";
-                cmd.Connection = sqlConnection1;
-
-                cmd.ExecuteNonQuery();
-                sqlConnection1.Close();*/
-                
-
-
-            }
+        }
 
         protected void Button3_Click(object sender, EventArgs e)
         {
